+++ src/main/scala/ru/tinkoff/gatling/profile/ProfileBuilder.scala
@@ -1,27 +1,8 @@
 package ru.tinkoff.gatling.profile
-import io.gatling.core.Predef._
-import io.gatling.core.structure.{ChainBuilder, ScenarioBuilder}
-import pureconfig.{ConfigReader, ConfigSource}
 import pureconfig.module.yaml.YamlConfigSource
+import pureconfig.{ConfigReader, ConfigSource}
 
 import scala.reflect.ClassTag
-
-trait RequestConfig {
-  val name: String
-  val url: String
-  val probability: Double
-
-  def toExec: ChainBuilder
-
-}
-
-trait ProfileConfig {
-  val name: String
-  protected val profile: Seq[RequestConfig]
-
-  def toRandomScenario(): Seq[(Double, ChainBuilder)] = profile.map(requestConfig => (requestConfig.probability, requestConfig.toExec))
-
-}
 
 class ProfileBuilder[A <: ProfileConfig] {
