+++ src/main/scala/ru/tinkoff/gatling/profile/RequestConfig.scala
@@ -1,0 +1,12 @@
+package ru.tinkoff.gatling.profile
+
+import io.gatling.core.structure.ChainBuilder
+
+trait RequestConfig {
+  val name: String
+  val url: String
+  val probability: Double
+
+  def toExec: ChainBuilder
+
+}
