+++ src/main/scala/ru/tinkoff/gatling/profile/http/HttpMethod.scala
@@ -1,0 +1,7 @@
+package ru.tinkoff.gatling.profile.http
+
+sealed trait HttpMethod
+case object GET extends HttpMethod
+case object POST extends HttpMethod
+case object PUT extends HttpMethod
+case object DELETE extends HttpMethod
